move rules to custom-rules folder
