+++ src/scanner/custom-rules/color-rule.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { RuleConfiguration } from './iruleresults';
+import { RuleConfiguration } from '../iruleresults';
 
 const colorCheckId: string = 'select-body';
