+++ src/scanner/custom-rules/css-positioning-rule.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import * as axe from 'axe-core';
-import { RuleConfiguration } from './iruleresults';
+import { RuleConfiguration } from '../iruleresults';
 
 const cssPositioningCheckId: string = 'css-positioning';
 const cssPositioningRuleId: string = 'css-positioning';
