+++ src/scanner/custom-rules/custom-widget.ts
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import * as AxeUtils from './axe-utils';
-import { generateARIACuesDictionary, generateHTMLCuesDictionary } from './cues';
-import { RuleConfiguration } from './iruleresults';
+import * as AxeUtils from '../axe-utils';
+import { generateARIACuesDictionary, generateHTMLCuesDictionary } from '../cues';
+import { RuleConfiguration } from '../iruleresults';
 
 const checkId = 'custom-widget';
 export const customWidgetConfiguration: RuleConfiguration = {
