+++ src/scanner/custom-rules/frame-title.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { IAxeCheckResultFrameExtraData, RuleConfiguration } from './iruleresults';
+import { IAxeCheckResultFrameExtraData, RuleConfiguration } from '../iruleresults';
 
 const frameTitleId: string = 'get-frame-title';
