+++ src/scanner/custom-rules/native-widgets-default.ts
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import * as AxeUtils from './axe-utils';
-import { RuleConfiguration } from './iruleresults';
-import { RoleUtils } from './role-utils';
+import * as AxeUtils from '../axe-utils';
+import { RuleConfiguration } from '../iruleresults';
+import { RoleUtils } from '../role-utils';
 
 export const nativeWidgetSelector = 'button, input[list], input[type]:not([type="hidden"]), select, textarea';
