+++ src/scanner/custom-rules/text-alternative.ts
@@ -1,9 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import * as AxeUtils from './axe-utils';
+import * as AxeUtils from '../axe-utils';
 
+import { RuleConfiguration } from '../iruleresults';
 import { isImage } from './image-rule';
-import { RuleConfiguration } from './iruleresults';
 
 export const textAlternativeConfiguration: RuleConfiguration = {
     checks: [
