+++ src/scanner/custom-rules/text-contrast.ts
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import * as AxeUtils from './axe-utils';
+import * as AxeUtils from '../axe-utils';
 
-import { RuleConfiguration } from './iruleresults';
+import { RuleConfiguration } from '../iruleresults';
 
 export const textContrastConfiguration: RuleConfiguration = {
     checks: [
