+++ src/scanner/custom-rules/widget-function.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import * as AxeUtils from './axe-utils';
-import { RuleConfiguration } from './iruleresults';
+import * as AxeUtils from '../axe-utils';
+import { RuleConfiguration } from '../iruleresults';
 import { createNativeWidgetConfiguration, getNativeWidgetElementType } from './native-widgets-default';
 
 export const widgetFunctionConfiguration: RuleConfiguration = createNativeWidgetConfiguration(
