+++ src/scanner/locale-configuration.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import * as Axe from 'axe-core';
-import { uniqueLandmarkRuleContent } from './unique-landmark';
+import { uniqueLandmarkRuleContent } from './custom-rules/unique-landmark';
 
 export const localeConfiguration: Axe.Locale = {
     rules: {
