+++ src/tests/unit/tests/scanner/helpers.ts
@@ -1,6 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { createNativeWidgetConfiguration, evaluateNativeWidget, nativeWidgetSelector } from '../../../../scanner/native-widgets-default';
+import {
+    createNativeWidgetConfiguration,
+    evaluateNativeWidget,
+    nativeWidgetSelector,
+} from '../../../../scanner/custom-rules/native-widgets-default';
 import { DictionaryStringTo } from '../../../../types/common-types';
 
 export function createNodeStub(tag: string, attributes: DictionaryStringTo<string>): HTMLElement {
