+++ src/tests/unit/tests/scanner/link-purpose.test.ts
@@ -3,7 +3,7 @@
 import { GlobalMock, GlobalScope, IMock, It, Mock, MockBehavior, Times } from 'typemoq';
 
 import * as AxeUtils from '../../../../scanner/axe-utils';
-import { linkPurposeConfiguration } from '../../../../scanner/link-purpose';
+import { linkPurposeConfiguration } from '../../../../scanner/custom-rules/link-purpose';
 
 describe('link purpose', () => {
     describe('verify link purpose configs', () => {
